requirements.txt
azure
boto
boto3
botocore
elasticsearch
fauxfactory
git+ssh://git@github.com/ansible/pytest-ansible.git
git+ssh://git@github.com/ansible/pytest-mp.git
git+ssh://git@github.com/jlaska/pytest-github.git
junit-xml
junit2html
junitparser
openshift
pip
pycodestyle
pycrypto
pytest
pytest-benchmark
pytest-ordering
pytest-rerunfailures
python-dateutil
pytz
requests
